requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.8.0
pymongo==4.6.0
python-dotenv==1.0.0
certifi==2023.11.17
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6

# LangChain and RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
pypdf==3.17.4

# LangGraph
langgraph==0.0.20

# Tavily for web search
tavily-python==0.3.0

# Google AI (Gemini)
google-generativeai==0.3.2